requirements.txt
streamlit
streamlit-autorefresh
pandas
Pillow
qrcode[pil]
reportlab
xlsxwriter